Refactor FinPlate component imports and update configurations
- Updated import paths for FinPlate and CleatAngle components to reflect new module structure.
- Adjusted configuration imports in FinPlate and FinPlateOutputDock components for consistency.
- Cleaned up code for improved readability and maintainability.
