+++ osdagclient/src/modules/shearConnection/endPlate/FinPlate.jsx
@@ -1,8 +1,8 @@
 import React from "react";
 import { EngineeringModule } from "../../shared/components/EngineeringModule";
-import { finPlateConfig } from "./configs/finPlateConfig";
 import FinPlateOutputDock from "./components/FinPlateOutputDock";
 import { menuItems } from "../../shared/utils/moduleUtils";
+import { finPlateConfig } from "../finPlate/configs/finPlateConfig";
 
 function FinPlate() {
   return (
