+++ osdagclient/src/modules/shearConnection/finPlate/FinPlate.jsx
@@ -1,9 +1,9 @@
 /* eslint-disable no-unused-vars */
 import React from "react";
 import { EngineeringModule } from "../../shared/components/EngineeringModule";
-import { finPlateConfig } from "./configs/finPlateConfig";
 import FinPlateOutputDock from "./components/FinPlateOutputDock";
 import { menuItems } from "../../shared/utils/moduleUtils";
+import { finPlateConfig } from "./configs/finPlateConfig";
 
 function FinPlate() {
   return (
